# backend/src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "AutoPwnDoc REST API",
    "description": "Interactive documentation for the AutoPwnDoc REST API. Authenticate requests with an API key in the `X-API-Key` header (or a logged-in browser session). New API keys act as the user who created them, including that user's current role and audit ownership. API keys can read taxonomy definitions and assign approved paths to audit findings, but cannot modify the taxonomy catalog. Standard JSON responses use `{ status, datas }`; file-download endpoints return a file instead.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header",
      "description": "Create and revoke API keys in Settings > API. Keys act as their creating user and can also be sent as a Bearer token."
    }
  },
  "paths": {
    "/api/audits": {
      "get": {
        "tags": [
          "Audit"
        ],
        "description": "",
        "parameters": [
          {
            "name": "findingTitle",
            "in": "query",
            "type": "string"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Audit"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "language": {
                  "example": "any"
                },
                "auditType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/data/vulnerability-taxonomy/hierarchy": {
      "get": {
        "tags": [
          "Data"
        ],
        "description": "Returns the approved taxonomy hierarchy in creation order. Use one exact path from this response when assigning taxonomies to audit findings.",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/mcp": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "map(handleMessage": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "202": {
            "description": "Accepted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/vulnerabilities": {
      "get": {
        "tags": [
          "Vulnerabilities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Vulnerabilities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "length": {
                  "example": "any"
                },
                "details": {
                  "example": "any"
                },
                "cvssv3": {
                  "example": "any"
                },
                "cvssv4": {
                  "example": "any"
                },
                "priority": {
                  "example": "any"
                },
                "remediationComplexity": {
                  "example": "any"
                },
                "taxonomies": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Vulnerabilities"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/ai/search-similar": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Search semantically similar vulnerabilities",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "description": "Search query and optional locale.",
            "schema": {
              "type": "object",
              "required": [
                "query"
              ],
              "properties": {
                "query": {
                  "type": "string"
                },
                "locale": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "403": {
            "description": "The requested AI capability is disabled or unavailable"
          },
          "422": {
            "description": "Invalid request parameters"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "required": {
          "type": "array",
          "example": [
            "status",
            "datas"
          ],
          "items": {
            "type": "string"
          }
        },
        "properties": {
          "type": "object",
          "properties": {
            "status": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "example": {
                  "type": "string",
                  "example": "success"
                }
              }
            },
            "datas": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "description": {
                  "type": "string",
                  "example": "Endpoint response payload."
                }
              }
            }
          }
        }
      }
    }
  }
}